Refactor I3C driver
