+++ src/i3c/ccc.rs
@@ -1,67 +1,70 @@
 // Licensed under the Apache-2.0 license
 
-use super::ast1060_i3c::HardwareInterface;
-use super::i3c_config::I3cConfig;
+//! I3C Common Command Codes (CCC)
+//!
+//! Functions and types for executing I3C CCCs.
 
-pub const I3C_CCC_RSTDAA: u8 = 0x6;
-pub const I3C_CCC_ENTDAA: u8 = 0x7;
-pub const I3C_CCC_SETHID: u8 = 0x61;
-pub const I3C_CCC_DEVCTRL: u8 = 0x62;
-pub const I3C_CCC_SETDASA: u8 = 0x87;
-pub const I3C_CCC_SETNEWDA: u8 = 0x88;
-pub const I3C_CCC_GETPID: u8 = 0x8d;
-pub const I3C_CCC_GETBCR: u8 = 0x8e;
-pub const I3C_CCC_GETSTATUS: u8 = 0x90;
+use super::config::I3cConfig;
+use super::constants::{
+    I3C_CCC_GETBCR, I3C_CCC_GETPID, I3C_CCC_GETSTATUS, I3C_CCC_RSTDAA, I3C_CCC_SETNEWDA,
+};
+use super::error::{CccErrorKind, I3cError};
+use super::hardware::HardwareInterface;
 
-pub const I3C_CCC_EVT_INTR: u8 = 1 << 0;
-pub const I3C_CCC_EVT_CR: u8 = 1 << 1;
-pub const I3C_CCC_EVT_HJ: u8 = 1 << 3;
-pub const I3C_CCC_EVT_ALL: u8 = I3C_CCC_EVT_INTR | I3C_CCC_EVT_CR | I3C_CCC_EVT_HJ;
+// =============================================================================
+// CCC Types
+// =============================================================================
 
-#[derive(Debug)]
-pub enum CccError {
-    InvalidParam,
-    NotFound,
-    NoFreeSlot,
-    Invalid,
-}
-
+/// CCC target payload for direct CCCs
 #[derive(Debug)]
 pub struct CccTargetPayload<'a> {
-    /// Target 7‑bit dynamic address (left‑aligned; driver decides if LSB is R/W).
+    /// Target 7-bit dynamic address
     pub addr: u8,
-    /// `false` = write, `true` = read.
+    /// `false` = write, `true` = read
     pub rnw: bool,
-    /// Data buffer for write (source) or read (destination).
+    /// Data buffer for write (source) or read (destination)
     pub data: Option<&'a mut [u8]>,
-    /// Actual bytes transferred (driver fills on return).
+    /// Actual bytes transferred (driver fills on return)
     pub num_xfer: usize,
 }
 
+/// CCC descriptor
 #[derive(Debug)]
 pub struct Ccc<'a> {
+    /// CCC ID (command code)
     pub id: u8,
-    /// Optional CCC data immediately following the CCC byte.
+    /// Optional CCC data immediately following the CCC byte
     pub data: Option<&'a mut [u8]>,
-    /// Actual bytes transferred (driver fills on return).
+    /// Actual bytes transferred (driver fills on return)
     pub num_xfer: usize,
 }
 
-/// One CCC transaction description.
+/// Complete CCC transaction description
 #[derive(Debug)]
 pub struct CccPayload<'a, 'b> {
+    /// The CCC command
     pub ccc: Option<Ccc<'a>>,
-    /// Optional list of direct‑CCC target payloads.
+    /// Optional list of direct-CCC target payloads
     pub targets: Option<&'b mut [CccTargetPayload<'a>]>,
 }
 
+// =============================================================================
+// CCC Reset Action
+// =============================================================================
+
+/// RSTACT defining byte values
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum CccRstActDefByte {
-    CccRstActNoReset = 0x0,
-    CccRstActPeriphralOnly = 0x1,
-    CccRstActResetWholeTarget = 0x2,
-    CccRstActDebugNetworkAdapter = 0x3,
-    CccRstActVirtualTargetDetect = 0x4,
+    /// No reset
+    NoReset = 0x0,
+    /// Reset peripheral only
+    PeriphralOnly = 0x1,
+    /// Reset whole target
+    ResetWholeTarget = 0x2,
+    /// Debug network adapter
+    DebugNetworkAdapter = 0x3,
+    /// Virtual target detect
+    VirtualTargetDetect = 0x4,
 }
 
 impl CccRstActDefByte {
@@ -71,12 +74,20 @@ impl CccRstActDefByte {
     }
 }
 
+// =============================================================================
+// GETSTATUS Types
+// =============================================================================
+
+/// GETSTATUS format selection
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum GetStatusFormat {
+    /// Format 1 (no defining byte)
     Fmt1,
+    /// Format 2 (with defining byte)
     Fmt2(GetStatusDefByte),
 }
 
+/// GETSTATUS defining byte values
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum GetStatusDefByte {
     /// 0x00 - TGTSTAT
@@ -89,24 +100,29 @@ impl GetStatusDefByte {
     #[inline]
     fn as_byte(self) -> u8 {
         match self {
-            GetStatusDefByte::TgtStat => 0x00,
-            GetStatusDefByte::Precr => 0x91,
+            Self::TgtStat => 0x00,
+            Self::Precr => 0x91,
         }
     }
 }
 
+/// GETSTATUS response
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum GetStatusResp {
-    Fmt1 {
-        status: u16,
-    },
+    /// Format 1 response
+    Fmt1 { status: u16 },
+    /// Format 2 response
     Fmt2 {
         kind: GetStatusDefByte,
         raw_u16: u16,
     },
 }
 
-const fn i3c_ccc_enec(broadcast: bool) -> u8 {
+// =============================================================================
+// CCC Helper Functions
+// =============================================================================
+
+const fn ccc_enec(broadcast: bool) -> u8 {
     if broadcast {
         0x00
     } else {
@@ -114,7 +130,7 @@ const fn i3c_ccc_enec(broadcast: bool) -> u8 {
     }
 }
 
-const fn i3c_ccc_disec(broadcast: bool) -> u8 {
+const fn ccc_disec(broadcast: bool) -> u8 {
     if broadcast {
         0x01
     } else {
@@ -122,7 +138,7 @@ const fn i3c_ccc_disec(broadcast: bool) -> u8 {
     }
 }
 
-const fn i3c_ccc_rstact(broadcast: bool) -> u8 {
+const fn ccc_rstact(broadcast: bool) -> u8 {
     if broadcast {
         0x2a
     } else {
@@ -130,21 +146,27 @@ const fn i3c_ccc_rstact(broadcast: bool) -> u8 {
     }
 }
 
+// =============================================================================
+// CCC Operations
+// =============================================================================
+
+/// Enable/disable events for all devices (broadcast)
 pub fn ccc_events_all_set<H>(
     hw: &mut H,
     config: &mut I3cConfig,
     enable: bool,
     events: u8,
-) -> Result<(), CccError>
+) -> Result<(), I3cError>
 where
     H: HardwareInterface,
 {
     let id = if enable {
-        i3c_ccc_enec(true)
+        ccc_enec(true)
     } else {
-        i3c_ccc_disec(true)
+        ccc_disec(true)
     };
-    let rc = hw.do_ccc(
+
+    hw.do_ccc(
         config,
         &mut CccPayload {
             ccc: Some(Ccc {
@@ -154,26 +176,23 @@ where
             }),
             targets: None,
         },
-    );
-
-    match rc {
-        Ok(()) => Ok(()),
-        _ => Err(CccError::Invalid),
-    }
+    )
+    .map_err(|_| I3cError::CccError(CccErrorKind::Invalid))
 }
 
+/// Enable/disable events for a specific device (direct)
 pub fn ccc_events_set<H>(
     hw: &mut H,
     config: &mut I3cConfig,
     da: u8,
     enable: bool,
     events: u8,
-) -> Result<(), CccError>
+) -> Result<(), I3cError>
 where
     H: HardwareInterface,
 {
     if da == 0 {
-        return Err(CccError::InvalidParam);
+        return Err(I3cError::CccError(CccErrorKind::InvalidParam));
     }
 
     let mut ev_buf = [events];
@@ -186,9 +205,9 @@ where
 
     let mut tgts = [tgt];
     let ccc_id = if enable {
-        i3c_ccc_enec(false)
+        ccc_enec(false)
     } else {
-        i3c_ccc_disec(false)
+        ccc_disec(false)
     };
     let ccc = Ccc {
         id: ccc_id,
@@ -201,24 +220,22 @@ where
         targets: Some(&mut tgts[..]),
     };
 
-    let rc = hw.do_ccc(config, &mut payload);
-    match rc {
-        Ok(()) => Ok(()),
-        _ => Err(CccError::Invalid),
-    }
+    hw.do_ccc(config, &mut payload)
+        .map_err(|_| I3cError::CccError(CccErrorKind::Invalid))
 }
 
+/// Execute RSTACT (Reset Action) broadcast
 pub fn ccc_rstact_all<H>(
     hw: &mut H,
     config: &mut I3cConfig,
     action: CccRstActDefByte,
-) -> Result<(), CccError>
+) -> Result<(), I3cError>
 where
     H: HardwareInterface,
 {
     let mut db = [action.as_byte()];
     let ccc = Ccc {
-        id: i3c_ccc_rstact(true),
+        id: ccc_rstact(true),
         data: Some(&mut db[..]),
         num_xfer: 0,
     };
@@ -227,19 +244,17 @@ where
         targets: None,
     };
 
-    let rc = hw.do_ccc(config, &mut payload);
-    match rc {
-        Ok(()) => Ok(()),
-        _ => Err(CccError::Invalid),
-    }
+    hw.do_ccc(config, &mut payload)
+        .map_err(|_| I3cError::CccError(CccErrorKind::Invalid))
 }
 
-pub fn ccc_getbcr<H>(hw: &mut H, config: &mut I3cConfig, dyn_addr: u8) -> Result<u8, CccError>
+/// Get BCR (Bus Characteristics Register) from a device
+pub fn ccc_getbcr<H>(hw: &mut H, config: &mut I3cConfig, dyn_addr: u8) -> Result<u8, I3cError>
 where
     H: HardwareInterface,
 {
     if dyn_addr == 0 {
-        return Err(CccError::InvalidParam);
+        return Err(I3cError::CccError(CccErrorKind::InvalidParam));
     }
 
     let mut bcr_buf = [0u8; 1];
@@ -262,34 +277,35 @@ where
         targets: Some(&mut tgts[..]),
     };
 
-    let rc = hw.do_ccc(config, &mut payload);
-    match rc {
-        Ok(()) => Ok(bcr_buf[0]),
-        _ => Err(CccError::Invalid),
-    }
+    hw.do_ccc(config, &mut payload)
+        .map_err(|_| I3cError::CccError(CccErrorKind::Invalid))?;
+
+    Ok(bcr_buf[0])
 }
 
+/// Set new dynamic address for a device
 pub fn ccc_setnewda<H>(
     hw: &mut H,
     config: &mut I3cConfig,
     curr_da: u8,
     new_da: u8,
-) -> Result<(), CccError>
+) -> Result<(), I3cError>
 where
     H: HardwareInterface,
 {
     if curr_da == 0 || new_da == 0 {
-        return Err(CccError::InvalidParam);
+        return Err(I3cError::CccError(CccErrorKind::InvalidParam));
     }
 
     let pos = config.attached.pos_of_addr(curr_da);
     if pos.is_none() {
-        return Err(CccError::NotFound);
+        return Err(I3cError::CccError(CccErrorKind::NotFound));
     }
 
     if !config.addrbook.is_free(new_da) {
-        return Err(CccError::NoFreeSlot);
+        return Err(I3cError::CccError(CccErrorKind::NoFreeSlot));
     }
+
     let mut new_dyn_addr = (new_da & 0x7F) << 1;
     let tgt = CccTargetPayload {
         addr: curr_da,
@@ -308,11 +324,8 @@ where
         targets: Some(&mut tgts[..]),
     };
 
-    let rc = hw.do_ccc(config, &mut payload);
-    match rc {
-        Ok(()) => Ok(()),
-        _ => Err(CccError::Invalid),
-    }
+    hw.do_ccc(config, &mut payload)
+        .map_err(|_| I3cError::CccError(CccErrorKind::Invalid))
 }
 
 fn bytes_to_pid(bytes: &[u8]) -> u64 {
@@ -322,7 +335,8 @@ fn bytes_to_pid(bytes: &[u8]) -> u64 {
         .fold(0u64, |acc, &b| (acc << 8) | u64::from(b))
 }
 
-pub fn ccc_getpid<H>(hw: &mut H, config: &mut I3cConfig, dyn_addr: u8) -> Result<u64, CccError>
+/// Get PID (Provisional ID) from a device
+pub fn ccc_getpid<H>(hw: &mut H, config: &mut I3cConfig, dyn_addr: u8) -> Result<u64, I3cError>
 where
     H: HardwareInterface,
 {
@@ -346,24 +360,23 @@ where
         targets: Some(&mut tgts[..]),
     };
 
-    let rc = hw.do_ccc(config, &mut payload);
-    match rc {
-        Ok(()) => Ok(bytes_to_pid(&pid_buf)),
-        _ => Err(CccError::Invalid),
-    }
+    hw.do_ccc(config, &mut payload)
+        .map_err(|_| I3cError::CccError(CccErrorKind::Invalid))?;
+
+    Ok(bytes_to_pid(&pid_buf))
 }
 
+/// Get status from a device
 pub fn ccc_getstatus<H>(
     hw: &mut H,
     config: &mut I3cConfig,
     da: u8,
     fmt: GetStatusFormat,
-) -> Result<GetStatusResp, CccError>
+) -> Result<GetStatusResp, I3cError>
 where
     H: HardwareInterface,
 {
     let mut data_buf = [0u8; 2];
-
     let mut defbyte_buf = [0u8; 1];
 
     let tgt = CccTargetPayload {
@@ -394,37 +407,35 @@ where
         targets: Some(&mut targets_arr[..]),
     };
 
-    let rc = hw.do_ccc(config, &mut payload);
-    match rc {
-        Ok(()) => {
-            let val = u16::from_be_bytes(data_buf);
+    hw.do_ccc(config, &mut payload)
+        .map_err(|_| I3cError::CccError(CccErrorKind::Invalid))?;
 
-            let resp = match kind_opt {
-                None => GetStatusResp::Fmt1 { status: val },
-                Some(kind) => GetStatusResp::Fmt2 { kind, raw_u16: val },
-            };
-            Ok(resp)
-        }
+    let val = u16::from_be_bytes(data_buf);
 
-        _ => Err(CccError::Invalid),
-    }
+    let resp = match kind_opt {
+        None => GetStatusResp::Fmt1 { status: val },
+        Some(kind) => GetStatusResp::Fmt2 { kind, raw_u16: val },
+    };
+    Ok(resp)
 }
 
-pub fn ccc_getstatus_fmt1<H>(hw: &mut H, config: &mut I3cConfig, da: u8) -> Result<u16, CccError>
+/// Get status (Format 1) from a device
+pub fn ccc_getstatus_fmt1<H>(hw: &mut H, config: &mut I3cConfig, da: u8) -> Result<u16, I3cError>
 where
     H: HardwareInterface,
 {
     match ccc_getstatus(hw, config, da, GetStatusFormat::Fmt1) {
         Ok(GetStatusResp::Fmt1 { status }) => Ok(status),
-        _ => Err(CccError::Invalid),
+        _ => Err(I3cError::CccError(CccErrorKind::Invalid)),
     }
 }
 
-pub fn ccc_rstdaa_all<H>(hw: &mut H, config: &mut I3cConfig) -> Result<(), CccError>
+/// Reset dynamic address assignment for all devices (broadcast)
+pub fn ccc_rstdaa_all<H>(hw: &mut H, config: &mut I3cConfig) -> Result<(), I3cError>
 where
     H: HardwareInterface,
 {
-    let rc = hw.do_ccc(
+    hw.do_ccc(
         config,
         &mut CccPayload {
             ccc: Some(Ccc {
@@ -434,9 +445,6 @@ where
             }),
             targets: None,
         },
-    );
-    match rc {
-        Ok(()) => Ok(()),
-        _ => Err(CccError::Invalid),
-    }
+    )
+    .map_err(|_| I3cError::CccError(CccErrorKind::Invalid))
 }
