+++ src/i3c/hardware.rs
@@ -1,205 +1,72 @@
 // Licensed under the Apache-2.0 license
 
-use crate::common::{DummyDelay, Logger};
-use crate::i3c::ccc::{
-    ccc_events_set, CccPayload, I3C_CCC_DEVCTRL, I3C_CCC_ENTDAA, I3C_CCC_EVT_INTR, I3C_CCC_SETHID,
-};
-use crate::i3c::i3c_config::{Completion, I3cConfig};
-use crate::i3c::ibi_workq;
+//! I3C Hardware Interface
+//!
+//! Defines the hardware abstraction traits and IRQ handling infrastructure.
+//!
+//! # Trait Hierarchy
+//!
+//! The hardware interface is split into focused sub-traits:
+//!
+//! ```text
+//! HardwareInterface (supertrait)
+//! ├── HardwareCore      - Init, IRQ, enable/disable
+//! ├── HardwareClock     - Clock configuration
+//! ├── HardwareFifo      - FIFO operations
+//! ├── HardwareTransfer  - Transfers, CCC, device management
+//! ├── HardwareRecovery  - SW mode, bus recovery
+//! └── HardwareTarget    - Target mode operations
+//! ```
+//!
+//! # Platform Initialization
+//!
+//! SCU operations (clock enable, reset control) are **not** part of these traits.
+//! They should be performed by the platform/board layer before creating the
+//! I3C controller.
+
 use core::cell::RefCell;
+use critical_section::Mutex;
+
+use super::ccc::{ccc_events_set, CccPayload};
+use super::config::I3cConfig;
+use super::constants::{
+    I3C_CCC_DEVCTRL, I3C_CCC_ENTDAA, I3C_CCC_EVT_INTR, I3C_CCC_SETHID,
+    bit, field_get, field_prep, CM_TFR_STS_MASTER_HALT, CM_TFR_STS_TARGET_HALT,
+    COMMAND_ATTR_ADDR_ASSGN_CMD, COMMAND_ATTR_SLAVE_DATA_CMD, COMMAND_ATTR_XFER_ARG,
+    COMMAND_ATTR_XFER_CMD, COMMAND_PORT_ARG_DATA_LEN, COMMAND_PORT_ARG_DB, COMMAND_PORT_ATTR,
+    COMMAND_PORT_CMD, COMMAND_PORT_CP, COMMAND_PORT_DBP, COMMAND_PORT_DEV_COUNT,
+    COMMAND_PORT_DEV_INDEX, COMMAND_PORT_READ_TRANSFER, COMMAND_PORT_ROC, COMMAND_PORT_SPEED,
+    COMMAND_PORT_TID, COMMAND_PORT_TOC, DEV_ADDR_TABLE_IBI_MDB, DEV_ADDR_TABLE_IBI_PEC,
+    DEV_ADDR_TABLE_SIR_REJECT, I3CG_REG1_SCL_IN_SW_MODE_EN, I3CG_REG1_SCL_IN_SW_MODE_VAL,
+    I3CG_REG1_SDA_IN_SW_MODE_EN, I3CG_REG1_SDA_IN_SW_MODE_VAL, I3C_BCR_IBI_PAYLOAD_HAS_DATA_BYTE,
+    I3C_BUS_I2C_FMP_TF_MAX_NS, I3C_BUS_I2C_FMP_THIGH_MIN_NS, I3C_BUS_I2C_FMP_TLOW_MIN_NS,
+    I3C_BUS_I2C_FMP_TR_MAX_NS, I3C_BUS_I2C_FM_TF_MAX_NS, I3C_BUS_I2C_FM_THIGH_MIN_NS,
+    I3C_BUS_I2C_FM_TLOW_MIN_NS, I3C_BUS_I2C_FM_TR_MAX_NS, I3C_BUS_I2C_STD_TF_MAX_NS,
+    I3C_BUS_I2C_STD_THIGH_MIN_NS, I3C_BUS_I2C_STD_TLOW_MIN_NS, I3C_BUS_I2C_STD_TR_MAX_NS,
+    I3C_BUS_THIGH_MAX_NS, I3C_MSG_READ, IBIQ_STATUS_IBI_DATA_LEN, IBIQ_STATUS_IBI_DATA_LEN_SHIFT,
+    IBIQ_STATUS_IBI_ID, IBIQ_STATUS_IBI_ID_SHIFT, INTR_CCC_UPDATED_STAT, INTR_DYN_ADDR_ASSGN_STAT,
+    INTR_IBI_THLD_STAT, INTR_RESP_READY_STAT, INTR_TRANSFER_ABORT_STAT, INTR_TRANSFER_ERR_STAT,
+    MAX_CMDS, NSEC_PER_SEC, RESET_CTRL_ALL, RESET_CTRL_QUEUES, RESET_CTRL_XFER_QUEUES,
+    RESPONSE_ERROR_IBA_NACK, RESPONSE_PORT_DATA_LEN_MASK, RESPONSE_PORT_DATA_LEN_SHIFT,
+    RESPONSE_PORT_ERR_STATUS_MASK, RESPONSE_PORT_ERR_STATUS_SHIFT, RESPONSE_PORT_TID_MASK,
+    RESPONSE_PORT_TID_SHIFT, SDA_TX_HOLD_MASK, SDA_TX_HOLD_MAX, SDA_TX_HOLD_MIN, SLV_DCR_MASK,
+    SLV_EVENT_CTRL_SIR_EN,
+};
+use super::error::I3cError as I3cDrvError;
+use super::error::I3cError;
+use super::ibi as ibi_workq;
+use super::types::{I3cCmd, I3cIbi, I3cMsg, I3cXfer, SpeedI3c, Tid};
+use crate::common::{DummyDelay, Logger};
+
 use core::marker::PhantomData;
 use core::ptr::read_volatile;
 use core::sync::atomic::Ordering;
 use cortex_m::peripheral::NVIC;
-use critical_section::Mutex;
 use embedded_hal::delay::DelayNs;
 
-pub const I3C_MSG_WRITE: u8 = 0x0;
-pub const I3C_MSG_READ: u8 = 0x1;
-pub const I3C_MSG_STOP: u8 = 0x2;
-
-pub const I3C_BUS_I2C_STD_TLOW_MIN_NS: u32 = 4_700;
-pub const I3C_BUS_I2C_STD_THIGH_MIN_NS: u32 = 4_000;
-pub const I3C_BUS_I2C_STD_TR_MAX_NS: u32 = 1_000;
-pub const I3C_BUS_I2C_STD_TF_MAX_NS: u32 = 300;
-
-pub const I3C_BUS_I2C_FM_TLOW_MIN_NS: u32 = 1_300;
-pub const I3C_BUS_I2C_FM_THIGH_MIN_NS: u32 = 600;
-pub const I3C_BUS_I2C_FM_TR_MAX_NS: u32 = 300;
-pub const I3C_BUS_I2C_FM_TF_MAX_NS: u32 = 300;
-
-pub const I3C_BUS_I2C_FMP_TLOW_MIN_NS: u32 = 500;
-pub const I3C_BUS_I2C_FMP_THIGH_MIN_NS: u32 = 260;
-pub const I3C_BUS_I2C_FMP_TR_MAX_NS: u32 = 120;
-pub const I3C_BUS_I2C_FMP_TF_MAX_NS: u32 = 120;
-
-pub const I3C_BUS_THIGH_MAX_NS: u32 = 41;
-
-pub const NSEC_PER_SEC: u32 = 1_000_000_000;
-pub const SDA_TX_HOLD_MIN: u32 = 0b001;
-pub const SDA_TX_HOLD_MAX: u32 = 0b111;
-pub const SDA_TX_HOLD_MASK: u32 = 0x0007_0000; // bits 18:16
-
-pub const SLV_DCR_MASK: u32 = 0x0000_ff00;
-pub const SLV_EVENT_CTRL: u32 = 0x38;
-pub const SLV_EVENT_CTRL_MWL_UPD: u32 = bit(7);
-pub const SLV_EVENT_CTRL_MRL_UPD: u32 = bit(6);
-pub const SLV_EVENT_CTRL_HJ_REQ: u32 = bit(3);
-pub const SLV_EVENT_CTRL_SIR_EN: u32 = bit(0);
-
-pub const I3CG_REG1_SCL_IN_SW_MODE_VAL: u32 = bit(23);
-pub const I3CG_REG1_SDA_IN_SW_MODE_VAL: u32 = bit(27);
-pub const I3CG_REG1_SCL_IN_SW_MODE_EN: u32 = bit(28);
-pub const I3CG_REG1_SDA_IN_SW_MODE_EN: u32 = bit(29);
-
-pub const CM_TFR_STS_MASTER_HALT: u8 = 0xf;
-pub const CM_TFR_STS_TARGET_HALT: u8 = 0x6;
-
-pub const COMMAND_QUEUE_PORT: u32 = 0x0c;
-
-// --- single-bit flags ---
-pub const COMMAND_PORT_PEC: u32 = bit(31);
-pub const COMMAND_PORT_TOC: u32 = bit(30);
-pub const COMMAND_PORT_READ_TRANSFER: u32 = bit(28);
-pub const COMMAND_PORT_SDAP: u32 = bit(27);
-pub const COMMAND_PORT_ROC: u32 = bit(26);
-pub const COMMAND_PORT_DBP: u32 = bit(25);
-pub const COMMAND_PORT_CP: u32 = bit(15);
-pub const COMMAND_PORT_SPEED: u32 = bits(23, 21);
-pub const COMMAND_PORT_DEV_INDEX: u32 = bits(20, 16);
-pub const COMMAND_PORT_CMD: u32 = bits(14, 7);
-pub const COMMAND_PORT_TID: u32 = bits(6, 3);
-pub const COMMAND_PORT_ARG_DB: u32 = bits(15, 8);
-pub const COMMAND_PORT_ARG_DATA_LEN: u32 = bits(31, 16);
-pub const COMMAND_PORT_ATTR: u32 = bits(2, 0);
-pub const COMMAND_PORT_DEV_COUNT: u32 = bits(25, 21);
-
-pub const TID_TARGET_IBI: u32 = 0x1;
-pub const TID_TARGET_RD_DATA: u32 = 0x2;
-pub const TID_TARGET_MASTER_WR: u32 = 0x8;
-pub const TID_TARGET_MASTER_DEF: u32 = 0xf;
-
-pub const COMMAND_ATTR_XFER_CMD: u32 = 0;
-pub const COMMAND_ATTR_XFER_ARG: u32 = 1;
-pub const COMMAND_ATTR_SHORT_ARG: u32 = 2;
-pub const COMMAND_ATTR_ADDR_ASSGN_CMD: u32 = 3;
-pub const COMMAND_ATTR_SLAVE_DATA_CMD: u32 = 0;
-
-pub const DEV_ADDR_TABLE_LEGACY_I2C_DEV: u32 = bit(31);
-pub const DEV_ADDR_TABLE_DYNAMIC_ADDR: u32 = bits(23, 16);
-pub const DEV_ADDR_TABLE_MR_REJECT: u32 = bit(14);
-pub const DEV_ADDR_TABLE_SIR_REJECT: u32 = bit(13);
-pub const DEV_ADDR_TABLE_IBI_MDB: u32 = bit(12);
-pub const DEV_ADDR_TABLE_IBI_PEC: u32 = bit(11);
-pub const DEV_ADDR_TABLE_STATIC_ADDR: u32 = bits(6, 0);
-
-pub const IBI_QUEUE_STATUS: u32 = 0x18;
-pub const IBIQ_STATUS_IBI_ID: u32 = bits(15, 8);
-pub const IBIQ_STATUS_IBI_ID_SHIFT: u32 = 8;
-pub const IBIQ_STATUS_IBI_DATA_LEN: u32 = bits(7, 0);
-pub const IBIQ_STATUS_IBI_DATA_LEN_SHIFT: u32 = 0;
-
-pub const RESET_CTRL_IBI_QUEUE: u32 = bit(5);
-pub const RESET_CTRL_RX_FIFO: u32 = bit(4);
-pub const RESET_CTRL_TX_FIFO: u32 = bit(3);
-pub const RESET_CTRL_RESP_QUEUE: u32 = bit(2);
-pub const RESET_CTRL_CMD_QUEUE: u32 = bit(1);
-pub const RESET_CTRL_SOFT: u32 = bit(0);
-
-pub const RESET_CTRL_ALL: u32 = RESET_CTRL_IBI_QUEUE
-    | RESET_CTRL_RX_FIFO
-    | RESET_CTRL_TX_FIFO
-    | RESET_CTRL_RESP_QUEUE
-    | RESET_CTRL_CMD_QUEUE
-    | RESET_CTRL_SOFT;
-
-pub const RESET_CTRL_QUEUES: u32 = RESET_CTRL_IBI_QUEUE
-    | RESET_CTRL_RX_FIFO
-    | RESET_CTRL_TX_FIFO
-    | RESET_CTRL_RESP_QUEUE
-    | RESET_CTRL_CMD_QUEUE;
-
-pub const RESET_CTRL_XFER_QUEUES: u32 =
-    RESET_CTRL_RX_FIFO | RESET_CTRL_TX_FIFO | RESET_CTRL_RESP_QUEUE | RESET_CTRL_CMD_QUEUE;
-
-pub const RESPONSE_QUEUE_PORT: u32 = 0x10;
-pub const RESPONSE_PORT_ERR_STATUS_SHIFT: u32 = 28;
-pub const RESPONSE_PORT_ERR_STATUS_MASK: u32 = genmask(31, 28);
-pub const RESPONSE_PORT_TID_SHIFT: u32 = 24;
-pub const RESPONSE_PORT_TID_MASK: u32 = genmask(27, 24);
-pub const RESPONSE_PORT_DATA_LEN_SHIFT: u32 = 0;
-pub const RESPONSE_PORT_DATA_LEN_MASK: u32 = genmask(15, 0);
-
-pub const RESPONSE_NO_ERROR: u32 = 0;
-pub const RESPONSE_ERROR_CRC: u32 = 1;
-pub const RESPONSE_ERROR_PARITY: u32 = 2;
-pub const RESPONSE_ERROR_FRAME: u32 = 3;
-pub const RESPONSE_ERROR_IBA_NACK: u32 = 4;
-pub const RESPONSE_ERROR_ADDRESS_NACK: u32 = 5;
-pub const RESPONSE_ERROR_OVER_UNDER_FLOW: u32 = 6;
-pub const RESPONSE_ERROR_TRANSF_ABORT: u32 = 8;
-pub const RESPONSE_ERROR_I2C_W_NACK_ERR: u32 = 9;
-pub const RESPONSE_ERROR_EARLY_TERMINATE: u32 = 10;
-pub const RESPONSE_ERROR_PEC_ERR: u32 = 12;
-
-pub const INTR_STATUS: u32 = 0x3c;
-pub const INTR_STATUS_EN: u32 = 0x40;
-pub const INTR_SIGNAL_EN: u32 = 0x44;
-pub const INTR_FORCE: u32 = 0x48;
-
-pub const INTR_BUSOWNER_UPDATE_STAT: u32 = bit(13);
-pub const INTR_IBI_UPDATED_STAT: u32 = bit(12);
-pub const INTR_READ_REQ_RECV_STAT: u32 = bit(11);
-pub const INTR_DEFSLV_STAT: u32 = bit(10);
-pub const INTR_TRANSFER_ERR_STAT: u32 = bit(9);
-pub const INTR_DYN_ADDR_ASSGN_STAT: u32 = bit(8);
-pub const INTR_CCC_UPDATED_STAT: u32 = bit(6);
-pub const INTR_TRANSFER_ABORT_STAT: u32 = bit(5);
-pub const INTR_RESP_READY_STAT: u32 = bit(4);
-pub const INTR_CMD_QUEUE_READY_STAT: u32 = bit(3);
-pub const INTR_IBI_THLD_STAT: u32 = bit(2);
-pub const INTR_RX_THLD_STAT: u32 = bit(1);
-pub const INTR_TX_THLD_STAT: u32 = bit(0);
-pub const I3C_BCR_IBI_PAYLOAD_HAS_DATA_BYTE: u32 = bit(2);
-
-#[must_use]
-pub const fn bit(n: u32) -> u32 {
-    1 << n
-}
-#[must_use]
-pub const fn bits(h: u32, l: u32) -> u32 {
-    ((1u32 << (h - l + 1)) - 1) << l
-}
-#[must_use]
-pub const fn field_prep(mask: u32, val: u32) -> u32 {
-    (val << mask.trailing_zeros()) & mask
-}
-const fn field_get(val: u32, mask: u32, shift: u32) -> u32 {
-    (val & mask) >> shift
-}
-const fn genmask(msb: u32, lsb: u32) -> u32 {
-    let width = msb - lsb + 1;
-    if width >= 32 {
-        u32::MAX
-    } else {
-        ((1u32 << width) - 1) << lsb
-    }
-}
-
-const MAX_CMDS: usize = 32;
-
-#[derive(Debug)]
-pub enum I3cDrvError {
-    NoDatPos,
-    NoMsgs,
-    TooManyMsgs,
-    InvalidArgs,
-    Timeout,
-    NoSuchDev,
-    Access,
-    IoError,
-    Invalid,
-}
+// =============================================================================
+// IRQ Handler Infrastructure
+// =============================================================================
 
 #[derive(Clone, Copy)]
 struct Handler {
@@ -214,6 +81,15 @@ static BUS_HANDLERS: [Mutex<RefCell<Option<Handler>>>; 4] = [
     Mutex::new(RefCell::new(None)),
 ];
 
+/// Register an IRQ handler for an I3C bus
+///
+/// # Arguments
+/// * `bus` - Bus index (0-3)
+/// * `func` - Handler function
+/// * `ctx` - Context value passed to handler
+///
+/// # Panics
+/// Panics if `bus >= 4`.
 pub fn register_i3c_irq_handler(bus: usize, func: fn(usize), ctx: usize) {
     assert!(bus < 4);
     critical_section::with(|cs| {
@@ -221,235 +97,248 @@ pub fn register_i3c_irq_handler(bus: usize, func: fn(usize), ctx: usize) {
     });
 }
 
+/// Dispatch IRQ for a specific bus
+///
+/// Called by the actual IRQ entry points (defined elsewhere to avoid symbol conflicts).
 #[inline]
-fn dispatch_irq(bus: usize) {
-    critical_section::with(|cs| {
-        if let Some(h) = *BUS_HANDLERS[bus].borrow(cs).borrow() {
-            (h.func)(h.ctx);
-        }
+pub fn dispatch_i3c_irq(bus: usize) {
+    // Copy handler out of critical section to avoid blocking IRQs during handler
+    let handler = critical_section::with(|cs| {
+        BUS_HANDLERS.get(bus)
+            .and_then(|m| *m.borrow(cs).borrow())
     });
-}
-
-#[no_mangle]
-pub extern "C" fn i3c() {
-    dispatch_irq(0);
-}
-#[no_mangle]
-pub extern "C" fn i3c1() {
-    dispatch_irq(1);
-}
-#[no_mangle]
-pub extern "C" fn i3c2() {
-    dispatch_irq(2);
-}
-#[no_mangle]
-pub extern "C" fn i3c3() {
-    dispatch_irq(3);
-}
-
-#[repr(u32)]
-pub enum SpeedI3c {
-    Sdr0 = 0x0,
-    Sdr1 = 0x1,
-    Sdr2 = 0x2,
-    Sdr3 = 0x3,
-    Sdr4 = 0x4,
-    HdrTs = 0x5,
-    HdrDdr = 0x6,
-    I2cFmAsI3c = 0x7,
-}
-
-#[repr(u32)]
-pub enum SpeedI2c {
-    Fm = 0x0,
-    Fmp = 0x1,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum Tid {
-    TargetIbi = 0x1,
-    TargetRdData = 0x2,
-    TargetMasterWr = 0x8,
-    TargetMasterDef = 0xF,
-}
-
-pub enum I3cStatus {
-    Ok,
-    Timeout,
-    Busy,
-    Pending,
-    Invalid,
-}
-
-#[derive(Debug)]
-pub struct I3cCmd<'a> {
-    pub cmd_lo: u32,
-    pub cmd_hi: u32,
-    pub tx: Option<&'a [u8]>,
-    pub rx: Option<&'a mut [u8]>,
-    pub tx_len: u32,
-    pub rx_len: u32,
-    pub ret: i32,
-}
-
-pub struct I3cMsg<'a> {
-    pub buf: Option<&'a mut [u8]>,
-    pub actual_len: u32,
-    pub num_xfer: u32,
-    pub flags: u8,
-    pub hdr_mode: u8,
-    pub hdr_cmd_mode: u8,
-}
-
-pub struct I3cXfer<'cmds, 'buf> {
-    pub cmds: &'cmds mut [I3cCmd<'buf>],
-    pub ret: i32,
-    pub done: Completion,
-}
-
-impl<'cmds, 'buf> I3cXfer<'cmds, 'buf> {
-    pub fn new(cmds: &'cmds mut [I3cCmd<'buf>]) -> Self {
-        Self {
-            cmds,
-            ret: 0,
-            done: Completion::new(),
-        }
-    }
-
-    pub fn ncmds(&self) -> usize {
-        self.cmds.len()
+    if let Some(h) = handler {
+        (h.func)(h.ctx);
     }
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct I3cPid(pub u64);
+// IRQ entry points - defined in src/i3c/ module to avoid symbol conflicts.
+// Use register_i3c_irq_handler() to register handlers that will be called
+// from those entry points.
 
-impl I3cPid {
-    #[must_use]
-    pub const fn manuf_id(self) -> u16 {
-        ((self.0 >> 33) & 0x1FFF) as u16
-    }
-    #[must_use]
-    pub const fn has_random_lower32(self) -> bool {
-        (self.0 & (1u64 << 32)) != 0
-    }
-}
+// =============================================================================
+// Sub-trait: Core Operations
+// =============================================================================
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct I3cDeviceId {
-    pub pid: I3cPid,
-}
-
-impl I3cDeviceId {
-    #[must_use]
-    pub const fn new(pid: u64) -> Self {
-        Self { pid: I3cPid(pid) }
-    }
-}
-
-pub const I3C_BROADCAST_ADDR: u8 = 0x7E;
-pub const I3C_MAX_ADDR: u8 = 0x7F;
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum I3cIbiType {
-    TargetIntr,
-    ControllerRoleRequest,
-    HotJoin,
-    WorkqueueCb,
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct I3cIbi<'a> {
-    pub ibi_type: I3cIbiType,
-    pub payload: Option<&'a [u8]>,
-}
-
-impl<'a> I3cIbi<'a> {
-    #[inline]
-    #[must_use]
-    pub fn payload_len(&self) -> u8 {
-        self.payload.map_or(0, |p| {
-            u8::try_from(p.len().min(u8::MAX as usize)).unwrap_or(u8::MAX)
-        })
-    }
-
-    #[must_use]
-    pub fn first_byte(&self) -> Option<u8> {
-        self.payload.and_then(|p| p.first().copied())
-    }
-}
-
-pub trait HardwareInterface {
+/// Core hardware operations: init, IRQ, enable/disable
+pub trait HardwareCore {
+    /// Initialize the I3C controller hardware
     fn init(&mut self, config: &mut I3cConfig);
+
+    /// Get the bus number for this instance
     fn bus_num(&self) -> u8;
+
+    /// Enable interrupts
     fn enable_irq(&mut self);
+
+    /// Disable interrupts
     fn disable_irq(&mut self);
+
+    /// Enable the I3C controller
     fn i3c_enable(&mut self, config: &I3cConfig);
+
+    /// Disable the I3C controller
     fn i3c_disable(&mut self, is_secondary: bool);
-    fn core_reset_assert(&mut self, bus: u8);
-    fn core_reset_deassert(&mut self, bus: u8);
-    fn global_reset_assert(&mut self);
-    fn global_reset_deassert(&mut self);
-    fn clock_on(&mut self, bus: u8);
+
+    /// Set the controller role (primary/secondary)
     fn set_role(&mut self, is_secondary: bool);
+
+    /// Main ISR handler
+    fn i3c_aspeed_isr(&mut self, config: &mut I3cConfig);
+}
+
+// =============================================================================
+// Sub-trait: Clock Configuration
+// =============================================================================
+
+/// Clock and timing configuration
+pub trait HardwareClock {
+    /// Initialize clock timing parameters
+    ///
+    /// Implementations should use `config.core_clk_hz` if set, falling back
+    /// to [`get_clock_rate()`](Self::get_clock_rate) for auto-detection.
     fn init_clock(&mut self, config: &mut I3cConfig);
-    fn get_clock_rate(&self) -> u32;
+
+
+    /// Calculate I2C clock dividers for given SCL frequency
     fn calc_i2c_clk(&mut self, fscl_hz: u32) -> (u32, u32);
-    fn init_pid(&mut self, config: &mut I3cConfig, bus: u8);
-    fn enter_sw_mode(&mut self);
-    fn exit_sw_mode(&mut self);
-    fn i3c_toggle_scl_in(&mut self, count: u32);
-    fn gen_internal_stop(&mut self);
-    fn even_parity(byte: u8) -> bool;
-    fn set_ibi_mdb(&mut self, mdb: u8);
-    fn exit_halt(&mut self, config: &mut I3cConfig);
-    fn enter_halt(&mut self, by_sw: bool, config: &mut I3cConfig);
-    fn reset_ctrl(&mut self, reset: u32);
+
+    /// Initialize the PID (Provisional ID) for this controller
+    fn init_pid(&mut self, config: &mut I3cConfig);
+}
+
+// =============================================================================
+// Sub-trait: FIFO Operations
+// =============================================================================
+
+/// FIFO read/write operations
+pub trait HardwareFifo {
+    /// Write to TX FIFO
     fn wr_tx_fifo(&mut self, bytes: &[u8]);
+
+    /// Read from FIFO using provided read function
     fn rd_fifo<F>(&mut self, read_word: F, out: &mut [u8])
     where
         F: FnMut() -> u32;
+
+    /// Drain FIFO without storing data
     fn drain_fifo<F>(&mut self, read_word: F, len: usize)
     where
         F: FnMut() -> u32;
+
+    /// Read from RX FIFO
     fn rd_rx_fifo(&mut self, out: &mut [u8]);
+
+    /// Read from IBI FIFO
     fn rd_ibi_fifo(&mut self, out: &mut [u8]);
-    fn ibi_enable(&mut self, config: &mut I3cConfig, addr: u8) -> Result<(), I3cDrvError>;
+}
+
+// =============================================================================
+// Sub-trait: Transfer Operations
+// =============================================================================
+
+/// Transfer, CCC, and device management operations
+pub trait HardwareTransfer {
+    /// Set the IBI Mandatory Data Byte
+    fn set_ibi_mdb(&mut self, mdb: u8);
+
+    /// Exit halt state
+    fn exit_halt(&mut self, config: &mut I3cConfig);
+
+    /// Enter halt state
+    fn enter_halt(&mut self, by_sw: bool, config: &mut I3cConfig);
+
+    /// Reset controller components (FIFOs, queues, etc.)
+    fn reset_ctrl(&mut self, reset: u32);
+
+    /// Enable IBI for a device
+    fn ibi_enable(&mut self, config: &mut I3cConfig, addr: u8) -> Result<(), I3cError>;
+
+    /// Start a transfer
     fn start_xfer(&mut self, config: &mut I3cConfig, xfer: &mut I3cXfer);
+
+    /// End a transfer
     fn end_xfer(&mut self, config: &mut I3cConfig);
+
+    /// Get DAT position for an address
     fn get_addr_pos(&mut self, config: &I3cConfig, addr: u8) -> Option<u8>;
+
+    /// Detach a device by DAT position
     fn detach_i3c_dev(&mut self, pos: usize);
-    fn attach_i3c_dev(&mut self, pos: usize, addr: u8) -> Result<(), I3cDrvError>;
-    fn do_ccc(&mut self, config: &mut I3cConfig, ccc: &mut CccPayload) -> Result<(), I3cDrvError>;
-    fn do_entdaa(&mut self, config: &mut I3cConfig, index: u32) -> Result<(), I3cDrvError>;
+
+    /// Attach a device to a DAT position
+    fn attach_i3c_dev(&mut self, pos: usize, addr: u8) -> Result<(), I3cError>;
+
+    /// Execute a CCC
+    fn do_ccc(&mut self, config: &mut I3cConfig, ccc: &mut CccPayload) -> Result<(), I3cError>;
+
+    /// Execute ENTDAA (Enter Dynamic Address Assignment)
+    fn do_entdaa(&mut self, config: &mut I3cConfig, index: u32) -> Result<(), I3cError>;
+
+    /// Build commands for private transfer
     fn priv_xfer_build_cmds<'a>(
         &mut self,
         cmds: &mut [I3cCmd<'a>],
         msgs: &mut [I3cMsg<'a>],
         pos: u8,
-    ) -> Result<(), I3cDrvError>;
+    ) -> Result<(), I3cError>;
+
+    /// Execute a private transfer
     fn priv_xfer(
         &mut self,
         config: &mut I3cConfig,
         pid: u64,
         msgs: &mut [I3cMsg],
-    ) -> Result<(), I3cDrvError>;
-    fn target_tx_write(&mut self, buf: &[u8]);
+    ) -> Result<(), I3cError>;
+
+    /// Handle IBI SIR (Slave Interrupt Request)
     fn handle_ibi_sir(&mut self, config: &mut I3cConfig, addr: u8, len: usize);
+
+    /// Handle all pending IBIs
     fn handle_ibis(&mut self, config: &mut I3cConfig);
-    fn i3c_aspeed_isr(&mut self, config: &mut I3cConfig);
-    // target apis
-    fn target_ibi_raise_hj(&self, config: &mut I3cConfig) -> Result<(), I3cDrvError>;
+}
+
+// =============================================================================
+// Sub-trait: Recovery / Software Mode
+// =============================================================================
+
+/// Software mode and bus recovery operations
+pub trait HardwareRecovery {
+    /// Enter software mode for manual bus control
+    fn enter_sw_mode(&mut self);
+
+    /// Exit software mode
+    fn exit_sw_mode(&mut self);
+
+    /// Toggle SCL line in software mode
+    fn i3c_toggle_scl_in(&mut self, count: u32);
+
+    /// Generate an internal STOP condition
+    fn gen_internal_stop(&mut self);
+
+    /// Calculate even parity for a byte
+    fn even_parity(byte: u8) -> bool;
+}
+
+// =============================================================================
+// Sub-trait: Target Mode Operations
+// =============================================================================
+
+/// Target (secondary) mode operations
+pub trait HardwareTarget {
+    /// Write data to target TX buffer
+    fn target_tx_write(&mut self, buf: &[u8]);
+
+    /// Raise a Hot-Join IBI (target mode)
+    fn target_ibi_raise_hj(&self, config: &mut I3cConfig) -> Result<(), I3cError>;
+
+    /// Handle response ready in target mode
     fn target_handle_response_ready(&mut self, config: &mut I3cConfig);
+
+    /// Notify pending read in target mode
     fn target_pending_read_notify(
         &mut self,
         config: &mut I3cConfig,
         buf: &[u8],
         notifier: &mut I3cIbi,
-    ) -> Result<(), I3cDrvError>;
+    ) -> Result<(), I3cError>;
+
+    /// Handle CCC update in target mode
     fn target_handle_ccc_update(&mut self, config: &mut I3cConfig);
 }
 
+// =============================================================================
+// Supertrait: Full Hardware Interface
+// =============================================================================
+
+/// Complete hardware abstraction for I3C controllers
+///
+/// This is a supertrait combining all sub-traits. Implementors must provide
+/// all operations.
+///
+/// # Sub-traits
+///
+/// - [`HardwareCore`] - Init, IRQ, enable/disable
+/// - [`HardwareClock`] - Clock configuration
+/// - [`HardwareFifo`] - FIFO operations
+/// - [`HardwareTransfer`] - Transfers, CCC, device management
+/// - [`HardwareRecovery`] - SW mode, bus recovery
+/// - [`HardwareTarget`] - Target mode operations
+pub trait HardwareInterface:
+    HardwareCore + HardwareClock + HardwareFifo + HardwareTransfer + HardwareRecovery + HardwareTarget
+{
+}
+
+// Blanket implementation: any type implementing all sub-traits implements HardwareInterface
+impl<T> HardwareInterface for T where
+    T: HardwareCore
+        + HardwareClock
+        + HardwareFifo
+        + HardwareTransfer
+        + HardwareRecovery
+        + HardwareTarget
+{
+}
 pub trait Instance {
     fn ptr() -> *const ast1060_pac::i3c::RegisterBlock;
     fn ptr_global() -> *const ast1060_pac::i3cglobal::RegisterBlock;
@@ -480,6 +369,25 @@ macro_i3c!(I3c, 0);
 macro_i3c!(I3c1, 1);
 macro_i3c!(I3c2, 2);
 macro_i3c!(I3c3, 3);
+
+// ISR Interrupt Handlers
+// TODO: Uncomment when replacing legacy driver
+ #[no_mangle]
+ pub extern "C" fn i3c() {
+     dispatch_i3c_irq(0);
+ }
+ #[no_mangle]
+ pub extern "C" fn i3c1 () {
+     dispatch_i3c_irq(1);
+ }
+ #[no_mangle]
+ pub extern "C" fn i3c2() {
+     dispatch_i3c_irq(2);
+ }
+ #[no_mangle]
+ pub extern "C" fn i3c3() {
+     dispatch_i3c_irq(3);
+ }
 
 pub struct Ast1060I3c<I3C: Instance, L: Logger> {
     pub i3c: &'static ast1060_pac::i3c::RegisterBlock,
@@ -513,19 +421,11 @@ macro_rules! i3c_debug {
     }};
 }
 
-#[allow(unused_macros)]
-macro_rules! read_i3cg_reg0 {
-    ($self:expr, $bus:expr) => {{
-        match $bus {
-            0 => $self.i3cg.i3c010().read().bits(),
-            1 => $self.i3cg.i3c020().read().bits(),
-            2 => $self.i3cg.i3c030().read().bits(),
-            3 => $self.i3cg.i3c040().read().bits(),
-            _ => panic!("invalid I3C bus index: {}", $bus),
-        }
-    }};
-}
+// -----------------------------------------------------------------------------
+// Register Helper Macros
+// -----------------------------------------------------------------------------
 
+#[allow(unused_macros)]
 macro_rules! read_i3cg_reg1 {
     ($self:expr, $bus:expr) => {{
         match $bus {
@@ -550,6 +450,18 @@ macro_rules! write_i3cg_reg0 {
     }};
 }
 
+macro_rules! read_i3cg_reg0 {
+    ($self:expr, $bus:expr) => {{
+        match $bus {
+            0 => $self.i3cg.i3c010().read().bits(),
+            1 => $self.i3cg.i3c020().read().bits(),
+            2 => $self.i3cg.i3c030().read().bits(),
+            3 => $self.i3cg.i3c040().read().bits(),
+            _ => panic!("invalid I3C bus index: {}", $bus),
+        }
+    }};
+}
+
 macro_rules! write_i3cg_reg1 {
     ($self:expr, $bus:expr, |$w:ident| $body:expr) => {{
         match $bus {
@@ -557,19 +469,6 @@ macro_rules! write_i3cg_reg1 {
             1 => $self.i3cg.i3c024().write(|$w| $body),
             2 => $self.i3cg.i3c034().write(|$w| $body),
             3 => $self.i3cg.i3c044().write(|$w| $body),
-            _ => panic!("invalid I3C bus index: {}", $bus),
-        }
-    }};
-}
-
-#[allow(unused_macros)]
-macro_rules! modify_i3cg_reg0 {
-    ($self:expr, $bus:expr, |$r:ident, $w:ident| $body:expr) => {{
-        match $bus {
-            0 => $self.i3cg.i3c010().modify(|$r, $w| $body),
-            1 => $self.i3cg.i3c020().modify(|$r, $w| $body),
-            2 => $self.i3cg.i3c030().modify(|$r, $w| $body),
-            3 => $self.i3cg.i3c040().modify(|$r, $w| $body),
             _ => panic!("invalid I3C bus index: {}", $bus),
         }
     }};
@@ -606,63 +505,14 @@ macro_rules! i3c_dat_read {
 macro_rules! i3c_dat_write {
     ($self:expr, $pos:expr, |$w:ident| $body:expr) => {{
         match ($pos) {
-            0 => {
-                $self.i3c.i3cd280().write(|$w| $body);
-            }
-            1 => {
-                $self.i3c.i3cd284().write(|$w| $body);
-            }
-            2 => {
-                $self.i3c.i3cd288().write(|$w| $body);
-            }
-            3 => {
-                $self.i3c.i3cd28c().write(|$w| $body);
-            }
-            4 => {
-                $self.i3c.i3cd290().write(|$w| $body);
-            }
-            5 => {
-                $self.i3c.i3cd294().write(|$w| $body);
-            }
-            6 => {
-                $self.i3c.i3cd298().write(|$w| $body);
-            }
-            7 => {
-                $self.i3c.i3cd29c().write(|$w| $body);
-            }
-            _ => { /* ignore */ }
-        }
-    }};
-}
-
-#[allow(unused_macros)]
-macro_rules! i3c_dat_modify {
-    ($self:expr, $pos:expr, |$r:ident, $w:ident| $body:expr) => {{
-        match ($pos) {
-            0 => {
-                $self.i3c.i3cd280().modify(|$r, $w| $body);
-            }
-            1 => {
-                $self.i3c.i3cd284().modify(|$r, $w| $body);
-            }
-            2 => {
-                $self.i3c.i3cd288().modify(|$r, $w| $body);
-            }
-            3 => {
-                $self.i3c.i3cd28c().modify(|$r, $w| $body);
-            }
-            4 => {
-                $self.i3c.i3cd290().modify(|$r, $w| $body);
-            }
-            5 => {
-                $self.i3c.i3cd294().modify(|$r, $w| $body);
-            }
-            6 => {
-                $self.i3c.i3cd298().modify(|$r, $w| $body);
-            }
-            7 => {
-                $self.i3c.i3cd29c().modify(|$r, $w| $body);
-            }
+            0 => { $self.i3c.i3cd280().write(|$w| $body); }
+            1 => { $self.i3c.i3cd284().write(|$w| $body); }
+            2 => { $self.i3c.i3cd288().write(|$w| $body); }
+            3 => { $self.i3c.i3cd28c().write(|$w| $body); }
+            4 => { $self.i3c.i3cd290().write(|$w| $body); }
+            5 => { $self.i3c.i3cd294().write(|$w| $body); }
+            6 => { $self.i3c.i3cd298().write(|$w| $body); }
+            7 => { $self.i3c.i3cd29c().write(|$w| $body); }
             _ => { /* ignore */ }
         }
     }};
@@ -695,12 +545,95 @@ where
     Err(PollError::Timeout)
 }
 
-impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
+impl<I3C: Instance, L: Logger> Ast1060I3c<I3C, L> {
+    fn _toggle_scl_in(&mut self, count: u32) {
+        let bus = I3C::BUS_NUM;
+        for _ in 0..count {
+            modify_i3cg_reg1!(self, bus, |r, w| unsafe {
+                w.bits(r.bits() & !I3CG_REG1_SCL_IN_SW_MODE_VAL)
+            });
+            modify_i3cg_reg1!(self, bus, |r, w| unsafe {
+                w.bits(r.bits() | I3CG_REG1_SCL_IN_SW_MODE_VAL)
+            });
+        }
+    }
+
+    fn _gen_internal_stop(&mut self) {
+        let bus = I3C::BUS_NUM;
+        modify_i3cg_reg1!(self, bus, |r, w| unsafe {
+            w.bits(r.bits() & !I3CG_REG1_SCL_IN_SW_MODE_VAL)
+        });
+        modify_i3cg_reg1!(self, bus, |r, w| unsafe {
+            w.bits(r.bits() & !I3CG_REG1_SDA_IN_SW_MODE_VAL)
+        });
+        modify_i3cg_reg1!(self, bus, |r, w| unsafe {
+            w.bits(r.bits() | I3CG_REG1_SCL_IN_SW_MODE_VAL)
+        });
+        modify_i3cg_reg1!(self, bus, |r, w| unsafe {
+            w.bits(r.bits() | I3CG_REG1_SDA_IN_SW_MODE_VAL)
+        });
+    }
+
+    fn _enter_sw_mode(&mut self) {
+        i3c_debug!(self.logger, "enter sw mode");
+        let bus = I3C::BUS_NUM;
+        let mut reg = read_i3cg_reg1!(self, bus);
+        reg |= I3CG_REG1_SCL_IN_SW_MODE_VAL | I3CG_REG1_SDA_IN_SW_MODE_VAL;
+        modify_i3cg_reg1!(self, bus, |_r, w| unsafe { w.bits(reg) });
+        reg |= I3CG_REG1_SCL_IN_SW_MODE_EN | I3CG_REG1_SDA_IN_SW_MODE_EN;
+        modify_i3cg_reg1!(self, bus, |_r, w| unsafe { w.bits(reg) });
+    }
+
+    fn _exit_sw_mode(&mut self) {
+        let bus = I3C::BUS_NUM;
+        let mut reg = read_i3cg_reg1!(self, bus);
+        reg &= !(I3CG_REG1_SCL_IN_SW_MODE_EN | I3CG_REG1_SDA_IN_SW_MODE_EN);
+        modify_i3cg_reg1!(self, bus, |_r, w| unsafe { w.bits(reg) });
+    }
+
+    fn core_reset_assert(&mut self, bus: u8) {
+        match bus {
+            0 => self.scu.scu050().modify(|_, w| w.rst_i3c0ctrl().set_bit()),
+            1 => self.scu.scu050().modify(|_, w| w.rst_i3c1ctrl().set_bit()),
+            2 => self.scu.scu050().modify(|_, w| w.rst_i3c2ctrl().set_bit()),
+            3 => self.scu.scu050().modify(|_, w| w.rst_i3c3ctrl().set_bit()),
+            _ => panic!("invalid I3C bus index: {bus}"),
+        };
+    }
+
+    fn core_reset_deassert(&mut self, bus: u8) {
+        let mask = 1u32 << (8 + u32::from(bus));
+        self.scu
+            .scu054()
+            .modify(|_, w| unsafe { w.scu050sys_rst_ctrl_clear_reg2().bits(mask) });
+    }
+
+    #[allow(dead_code)]
+    fn global_reset_assert(&mut self) {
+        self.scu
+            .scu050()
+            .modify(|_, w| w.rst_i3cregdmactrl().set_bit());
+    }
+
+    fn global_reset_deassert(&mut self) {
+        self.scu
+            .scu054()
+            .modify(|_, w| unsafe { w.scu050sys_rst_ctrl_clear_reg2().bits(0x80) });
+    }
+
+    fn clock_on(&mut self, bus: u8) {
+        let mask = 1u32 << (8 + u32::from(bus));
+        self.scu
+            .scu094()
+            .modify(|_, w| unsafe { w.scu090clk_stop_ctrl_clear_reg_set2().bits(mask) });
+    }
+}
+
+impl<I3C: Instance, L: Logger> HardwareCore for Ast1060I3c<I3C, L> {
     #[allow(clippy::too_many_lines)]
     fn init(&mut self, config: &mut I3cConfig) {
         i3c_debug!(self.logger, "i3c init");
 
-        // Global reset is shared, so just need to deassert it
         self.global_reset_deassert();
 
         write_i3cg_reg1!(self, I3C::BUS_NUM, |w| unsafe {
@@ -741,7 +674,7 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
             I3C::BUS_NUM,
             config.is_secondary
         );
-        // Reset controller
+        
         self.i3c.i3cd034().write(|w| {
             w.ibiqueue_sw_rst()
                 .set_bit()
@@ -767,7 +700,7 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
 
         self.set_role(config.is_secondary);
         self.init_clock(config);
-        // init interrupt mask
+        
         self.i3c.i3cd03c().write(|w| unsafe { w.bits(0xffff_ffff) });
         if config.is_secondary {
             self.i3c.i3cd040().write(|w| {
@@ -817,7 +750,6 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
 
         config.sir_allowed_by_sw = false;
 
-        // Init hardware queues
         self.i3c
             .i3cd01c()
             .write(|w| unsafe { w.ibidata_threshold_value().bits(31) });
@@ -826,10 +758,8 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
             .i3cd020()
             .modify(|_, w| unsafe { w.rx_buffer_threshold_value().bits(0) });
 
-        // Init PID and DCR for target/secondary mode
-        self.init_pid(config, I3C::BUS_NUM);
+        self.init_pid(config);
 
-        // Get max device and DAT start addr
         config.maxdevs = self.i3c.i3cd05c().read().devaddrtabledepth().bits();
         config.free_pos = if config.maxdevs == 32 {
             u32::MAX
@@ -838,7 +768,6 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         };
         config.need_da = 0;
 
-        // Init DAT
         for i in 0..(config.maxdevs) {
             i3c_dat_write!(self, i, |w| {
                 w.sirreject().set_bit().mrreject().set_bit()
@@ -864,7 +793,6 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         self.i3c_enable(config);
 
         i3c_debug!(self.logger, "i3c enabled");
-        // Enable hot-join
         if !config.is_secondary {
             self.i3c
                 .i3cd040()
@@ -877,11 +805,9 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
             .i3cd000()
             .modify(|_, w| w.hot_join_ack_nack_ctrl().clear_bit());
         i3c_debug!(self.logger, "i3c init done");
-        i3c_debug!(
-            self.logger,
-            "i3c i3cd000: {:#x}",
-            self.i3c.i3cd000().read().bits()
-        );
+        
+        // Safety: Ensure memory barrier and init completion before interrupts are enabled by the caller
+        core::sync::atomic::compiler_fence(Ordering::SeqCst);
     }
 
     fn bus_num(&self) -> u8 {
@@ -917,54 +843,50 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         }
 
         if is_secondary {
-            // enter sw mode
-            self.enter_sw_mode();
+            self._enter_sw_mode();
         }
         self.i3c
             .i3cd000()
             .modify(|_, w| w.enbl_i3cctrl().clear_bit());
 
         if is_secondary {
-            self.i3c_toggle_scl_in(8);
-            self.gen_internal_stop();
-            self.exit_sw_mode();
+            self._toggle_scl_in(8);
+            self._gen_internal_stop();
+            self._exit_sw_mode();
         }
     }
 
-    fn core_reset_assert(&mut self, bus: u8) {
-        match bus {
-            0 => self.scu.scu050().modify(|_, w| w.rst_i3c0ctrl().set_bit()),
-            1 => self.scu.scu050().modify(|_, w| w.rst_i3c1ctrl().set_bit()),
-            2 => self.scu.scu050().modify(|_, w| w.rst_i3c2ctrl().set_bit()),
-            3 => self.scu.scu050().modify(|_, w| w.rst_i3c3ctrl().set_bit()),
-            _ => panic!("invalid I3C bus index: {bus}"),
-        };
-    }
-
-    fn core_reset_deassert(&mut self, bus: u8) {
-        let mask = 1u32 << (8 + u32::from(bus));
-        self.scu
-            .scu054()
-            .modify(|_, w| unsafe { w.scu050sys_rst_ctrl_clear_reg2().bits(mask) });
-    }
-
-    fn global_reset_assert(&mut self) {
-        self.scu
-            .scu050()
-            .modify(|_, w| w.rst_i3cregdmactrl().set_bit());
-    }
-
-    fn global_reset_deassert(&mut self) {
-        self.scu
-            .scu054()
-            .modify(|_, w| unsafe { w.scu050sys_rst_ctrl_clear_reg2().bits(0x80) });
-    }
-
-    fn clock_on(&mut self, bus: u8) {
-        let mask = 1u32 << (8 + u32::from(bus));
-        self.scu
-            .scu094()
-            .modify(|_, w| unsafe { w.scu090clk_stop_ctrl_clear_reg_set2().bits(mask) });
+    fn i3c_enable(&mut self, config: &I3cConfig) {
+        i3c_debug!(self.logger, "i3c enable");
+        if config.is_secondary {
+            i3c_debug!(self.logger, "i3c enable as secondary");
+            self.i3c.i3cd038().write(|w| unsafe { w.bits(0) });
+            self._enter_sw_mode();
+            self.i3c.i3cd000().modify(|_, w| {
+                w.enbl_adaption_of_i2ci3cmode()
+                    .clear_bit()
+                    .ibipayloaden()
+                    .set_bit()
+                    .enbl_i3cctrl()
+                    .set_bit()
+            });
+            let wait_cnt = &self.i3c.i3cd0d4().read().i3cibifree().bits();
+            let wait_ns = u32::from(*wait_cnt) * config.core_period;
+            let mut delay = DummyDelay {};
+            delay.delay_ns(wait_ns * 100_u32);
+            self._toggle_scl_in(8);
+            if self.i3c.i3cd000().read().enbl_i3cctrl().bit_is_set() {
+                self._gen_internal_stop();
+            }
+            self._exit_sw_mode();
+        } else {
+            self.i3c.i3cd000().modify(|_, w| {
+                w.i3cbroadcast_addr_include()
+                    .set_bit()
+                    .enbl_i3cctrl()
+                    .set_bit()
+            });
+        }
     }
 
     fn set_role(&mut self, is_secondary: bool) {
@@ -979,8 +901,51 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         }
     }
 
+    fn i3c_aspeed_isr(&mut self, config: &mut I3cConfig) {
+        self.disable_irq();
+        let status = self.i3c.i3cd03c().read().bits();
+        i3c_debug!(self.logger, "[ISR] 0x{:08x}", status);
+        if status == 0 {
+            self.enable_irq();
+            return;
+        }
+
+        if config.is_secondary {
+            if status & INTR_DYN_ADDR_ASSGN_STAT != 0 {
+                let da = self.i3c.i3cd004().read().dev_dynamic_addr().bits();
+                if let Some(tc) = &mut config.target_config {
+                    tc.addr = Some(da);
+                }
+                ibi_workq::i3c_ibi_work_enqueue_target_da_assignment(I3C::BUS_NUM.into());
+            }
+
+            if (status & INTR_RESP_READY_STAT) != 0 {
+                self.target_handle_response_ready(config);
+            }
+
+            if (status & INTR_CCC_UPDATED_STAT) != 0 {
+                self.target_handle_ccc_update(config);
+            }
+        } else {
+            if (status & (INTR_RESP_READY_STAT | INTR_TRANSFER_ERR_STAT | INTR_TRANSFER_ABORT_STAT))
+                != 0
+            {
+                self.end_xfer(config);
+            }
+
+            if (status & INTR_IBI_THLD_STAT) != 0 {
+                self.handle_ibis(config);
+            }
+        }
+
+        self.i3c.i3cd03c().write(|w| unsafe { w.bits(status) });
+        self.enable_irq();
+    }
+}
+
+impl<I3C: Instance, L: Logger> HardwareClock for Ast1060I3c<I3C, L> {
     fn init_clock(&mut self, config: &mut I3cConfig) {
-        let clk_rate = self.get_clock_rate();
+        let clk_rate = config.core_clk_hz.expect("core_clk_hz must be configured");
         i3c_debug!(self.logger, "i3c clock rate: {} Hz", clk_rate);
         config.core_period = (NSEC_PER_SEC).div_ceil(clk_rate);
 
@@ -1039,149 +1004,143 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
             };
         self.i3c.i3cd0b8().write(|w| unsafe {
             w.i3cpphcnt()
-                .bits(ns_to_cnt_u8(i3c_pp_hi_ns))
-                .i3cpplcnt()
-                .bits(ns_to_cnt_u8(i3c_pp_lo_ns))
-        });
+            .bits(ns_to_cnt_u8(i3c_pp_hi_ns))
+            .i3cpplcnt()
+            .bits(ns_to_cnt_u8(i3c_pp_lo_ns))
+    });
 
-        // SDA TX hold time
-        let hold_steps = (config.sda_tx_hold_ns)
-            .div_ceil(config.core_period)
-            .clamp(SDA_TX_HOLD_MIN, SDA_TX_HOLD_MAX);
-        let mut reg = self.i3c.i3cd0d0().read().bits();
-        reg = (reg & !SDA_TX_HOLD_MASK) | ((hold_steps & 0x7) << 16);
-        self.i3c.i3cd0d0().write(|w| unsafe { w.bits(reg) });
+    // SDA TX hold time
+    let hold_steps = (config.sda_tx_hold_ns)
+        .div_ceil(config.core_period)
+        .clamp(SDA_TX_HOLD_MIN, SDA_TX_HOLD_MAX);
+    let mut reg = self.i3c.i3cd0d0().read().bits();
+    reg = (reg & !SDA_TX_HOLD_MASK) | ((hold_steps & 0x7) << 16);
+    self.i3c.i3cd0d0().write(|w| unsafe { w.bits(reg) });
 
-        // BUS_FREE_TIMING
-        self.i3c.i3cd0d4().write(|w| unsafe { w.bits(0xffff_007c) });
+    // BUS_FREE_TIMING
+    self.i3c.i3cd0d4().write(|w| unsafe { w.bits(0xffff_007c) });
+}
+
+fn calc_i2c_clk(&mut self, fscl_hz: u32) -> (u32, u32) {
+    use core::cmp::max;
+
+    debug_assert!(fscl_hz > 0);
+    let period_ns: u32 = (1_000_000_000u32).div_ceil(fscl_hz);
+
+    let (lo_min, hi_min): (u32, u32) = if fscl_hz <= 100_000 {
+        (
+            (I3C_BUS_I2C_STD_TLOW_MIN_NS + I3C_BUS_I2C_STD_TF_MAX_NS).div_ceil(period_ns),
+            (I3C_BUS_I2C_STD_THIGH_MIN_NS + I3C_BUS_I2C_STD_TR_MAX_NS).div_ceil(period_ns),
+        )
+    } else if fscl_hz <= 400_000 {
+        (
+            (I3C_BUS_I2C_FM_TLOW_MIN_NS + I3C_BUS_I2C_FM_TF_MAX_NS).div_ceil(period_ns),
+            (I3C_BUS_I2C_FM_THIGH_MIN_NS + I3C_BUS_I2C_FM_TR_MAX_NS).div_ceil(period_ns),
+        )
+    } else {
+        (
+            (I3C_BUS_I2C_FMP_TLOW_MIN_NS + I3C_BUS_I2C_FMP_TF_MAX_NS).div_ceil(period_ns),
+            (I3C_BUS_I2C_FMP_THIGH_MIN_NS + I3C_BUS_I2C_FMP_TR_MAX_NS).div_ceil(period_ns),
+        )
+    };
+
+    let leftover = period_ns.saturating_sub(lo_min + hi_min);
+    let lo = lo_min + leftover / 2;
+    let hi = max(period_ns.saturating_sub(lo), hi_min);
+
+    (hi as u32, lo)
+}
+
+fn init_pid(&mut self, config: &mut I3cConfig) {
+    let bus = I3C::BUS_NUM;
+    self.i3c
+        .i3cd070()
+        .write(|w| unsafe { w.slvmipimfgid().bits(0x3f6).slvpiddcr().clear_bit() });
+
+    let rev_id: u32 = self.scu.scu004().read().hw_rev_id().bits().into();
+    let mut reg: u32 = rev_id << 16 | u32::from(bus) << 12;
+    reg |= 0xa000_0000;
+    self.i3c.i3cd074().write(|w| unsafe { w.bits(reg) });
+    let mut reg: u32 = self.i3c.i3cd078().read().bits();
+    reg &= !SLV_DCR_MASK;
+    reg |= (config.dcr << 8) | 0x66;
+    self.i3c.i3cd078().write(|w| unsafe { w.bits(reg) });
+}
+}
+
+impl<I3C: Instance, L: Logger> HardwareFifo for Ast1060I3c<I3C, L> {
+    fn wr_tx_fifo(&mut self, bytes: &[u8]) {
+        let mut chunks = bytes.chunks_exact(4);
+        for chunk in &mut chunks {
+            let word = u32::from_le_bytes([chunk[0], chunk[1], chunk[2], chunk[3]]);
+            self.i3c
+                .i3cd014()
+                .write(|w| unsafe { w.tx_data_port().bits(word) });
+        }
+
+        let rem = chunks.remainder();
+        if !rem.is_empty() {
+            let mut tmp = [0u8; 4];
+            tmp[..rem.len()].copy_from_slice(rem);
+            let word = u32::from_le_bytes(tmp);
+            self.i3c
+                .i3cd014()
+                .write(|w| unsafe { w.tx_data_port().bits(word) });
+        }
     }
 
-    fn get_clock_rate(&self) -> u32 {
-        200_000_000
+    fn rd_fifo<F>(&mut self, mut read_word: F, out: &mut [u8])
+    where
+        F: FnMut() -> u32,
+    {
+        let mut chunks = out.chunks_exact_mut(4);
+        for chunk in &mut chunks {
+            let val = read_word();
+            chunk.copy_from_slice(&val.to_le_bytes());
+        }
+
+        let rem = chunks.into_remainder();
+        if !rem.is_empty() {
+            let val = read_word();
+            let bytes = val.to_le_bytes();
+            rem.copy_from_slice(&bytes[..rem.len()]);
+        }
     }
 
-    fn calc_i2c_clk(&mut self, fscl_hz: u32) -> (u32, u32) {
-        use core::cmp::max;
-
-        debug_assert!(fscl_hz > 0);
-        let period_ns: u32 = (1_000_000_000u32).div_ceil(fscl_hz);
-
-        let (lo_min, hi_min): (u32, u32) = if fscl_hz <= 100_000 {
-            (
-                (I3C_BUS_I2C_STD_TLOW_MIN_NS + I3C_BUS_I2C_STD_TF_MAX_NS).div_ceil(period_ns),
-                (I3C_BUS_I2C_STD_THIGH_MIN_NS + I3C_BUS_I2C_STD_TR_MAX_NS).div_ceil(period_ns),
-            )
-        } else if fscl_hz <= 400_000 {
-            (
-                (I3C_BUS_I2C_FM_TLOW_MIN_NS + I3C_BUS_I2C_FM_TF_MAX_NS).div_ceil(period_ns),
-                (I3C_BUS_I2C_FM_THIGH_MIN_NS + I3C_BUS_I2C_FM_TR_MAX_NS).div_ceil(period_ns),
-            )
-        } else {
-            (
-                (I3C_BUS_I2C_FMP_TLOW_MIN_NS + I3C_BUS_I2C_FMP_TF_MAX_NS).div_ceil(period_ns),
-                (I3C_BUS_I2C_FMP_THIGH_MIN_NS + I3C_BUS_I2C_FMP_TR_MAX_NS).div_ceil(period_ns),
-            )
-        };
-
-        let leftover = period_ns.saturating_sub(lo_min + hi_min);
-        let lo = lo_min + leftover / 2;
-        let hi = max(period_ns.saturating_sub(lo), hi_min);
-
-        (hi as u32, lo)
+    fn drain_fifo<F>(&mut self, mut read_word: F, len: usize)
+    where
+        F: FnMut() -> u32,
+    {
+        let nwords = (len + 3) >> 2;
+        for _ in 0..nwords {
+            let _ = read_word();
+        }
     }
 
-    fn init_pid(&mut self, config: &mut I3cConfig, bus: u8) {
-        self.i3c
-            .i3cd070()
-            .write(|w| unsafe { w.slvmipimfgid().bits(0x3f6).slvpiddcr().clear_bit() });
-
-        let rev_id: u32 = self.scu.scu004().read().hw_rev_id().bits().into();
-        let mut reg: u32 = rev_id << 16 | u32::from(bus) << 12;
-        reg |= 0xa000_0000;
-        self.i3c.i3cd074().write(|w| unsafe { w.bits(reg) });
-        let mut reg: u32 = self.i3c.i3cd078().read().bits();
-        reg &= !SLV_DCR_MASK;
-        reg |= (config.dcr << 8) | 0x66;
-        self.i3c.i3cd078().write(|w| unsafe { w.bits(reg) });
+    fn rd_rx_fifo(&mut self, out: &mut [u8]) {
+        self.rd_fifo(|| self.i3c.i3cd014().read().rx_data_port().bits(), out);
     }
 
+    fn rd_ibi_fifo(&mut self, out: &mut [u8]) {
+        self.rd_fifo(|| self.i3c.i3cd018().read().bits(), out);
+    }
+}
+
+impl<I3C: Instance, L: Logger> HardwareRecovery for Ast1060I3c<I3C, L> {
     fn enter_sw_mode(&mut self) {
-        i3c_debug!(self.logger, "enter sw mode");
-        let bus = I3C::BUS_NUM;
-        let mut reg = read_i3cg_reg1!(self, bus);
-        reg |= I3CG_REG1_SCL_IN_SW_MODE_VAL | I3CG_REG1_SDA_IN_SW_MODE_VAL;
-        modify_i3cg_reg1!(self, bus, |_r, w| unsafe { w.bits(reg) });
-        reg |= I3CG_REG1_SCL_IN_SW_MODE_EN | I3CG_REG1_SDA_IN_SW_MODE_EN;
-        modify_i3cg_reg1!(self, bus, |_r, w| unsafe { w.bits(reg) });
+        self._enter_sw_mode();
     }
 
     fn exit_sw_mode(&mut self) {
-        let bus = I3C::BUS_NUM;
-        let mut reg = read_i3cg_reg1!(self, bus);
-        reg &= !(I3CG_REG1_SCL_IN_SW_MODE_EN | I3CG_REG1_SDA_IN_SW_MODE_EN);
-        modify_i3cg_reg1!(self, bus, |_r, w| unsafe { w.bits(reg) });
-    }
-
-    fn i3c_enable(&mut self, config: &I3cConfig) {
-        i3c_debug!(self.logger, "i3c enable");
-        if config.is_secondary {
-            i3c_debug!(self.logger, "i3c enable as secondary");
-            self.i3c.i3cd038().write(|w| unsafe { w.bits(0) });
-            self.enter_sw_mode();
-            // Enable hot-join
-            self.i3c.i3cd000().modify(|_, w| {
-                w.enbl_adaption_of_i2ci3cmode()
-                    .clear_bit()
-                    .ibipayloaden()
-                    .set_bit()
-                    .enbl_i3cctrl()
-                    .set_bit()
-            });
-            let wait_cnt = &self.i3c.i3cd0d4().read().i3cibifree().bits();
-            let wait_ns = u32::from(*wait_cnt) * config.core_period;
-            let mut delay = DummyDelay {};
-            delay.delay_ns(wait_ns * 100_u32);
-            self.i3c_toggle_scl_in(8);
-            if self.i3c.i3cd000().read().enbl_i3cctrl().bit_is_set() {
-                self.gen_internal_stop();
-            }
-            self.exit_sw_mode();
-        } else {
-            self.i3c.i3cd000().modify(|_, w| {
-                w.i3cbroadcast_addr_include()
-                    .set_bit()
-                    .enbl_i3cctrl()
-                    .set_bit()
-            });
-        }
+        self._exit_sw_mode();
     }
 
     fn i3c_toggle_scl_in(&mut self, count: u32) {
-        let bus = I3C::BUS_NUM;
-        for _ in 0..count {
-            modify_i3cg_reg1!(self, bus, |r, w| unsafe {
-                w.bits(r.bits() & !I3CG_REG1_SCL_IN_SW_MODE_VAL)
-            });
-            modify_i3cg_reg1!(self, bus, |r, w| unsafe {
-                w.bits(r.bits() | I3CG_REG1_SCL_IN_SW_MODE_VAL)
-            });
-        }
+        self._toggle_scl_in(count);
     }
 
     fn gen_internal_stop(&mut self) {
-        let bus = I3C::BUS_NUM;
-        modify_i3cg_reg1!(self, bus, |r, w| unsafe {
-            w.bits(r.bits() & !I3CG_REG1_SCL_IN_SW_MODE_VAL)
-        });
-        modify_i3cg_reg1!(self, bus, |r, w| unsafe {
-            w.bits(r.bits() & !I3CG_REG1_SDA_IN_SW_MODE_VAL)
-        });
-        modify_i3cg_reg1!(self, bus, |r, w| unsafe {
-            w.bits(r.bits() | I3CG_REG1_SCL_IN_SW_MODE_VAL)
-        });
-        modify_i3cg_reg1!(self, bus, |r, w| unsafe {
-            w.bits(r.bits() | I3CG_REG1_SDA_IN_SW_MODE_VAL)
-        });
+        self._gen_internal_stop();
     }
 
     fn even_parity(byte: u8) -> bool {
@@ -1195,7 +1154,9 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
 
         !parity
     }
+}
 
+impl<I3C: Instance, L: Logger> HardwareTransfer for Ast1060I3c<I3C, L> {
     fn set_ibi_mdb(&mut self, mdb: u8) {
         self.i3c
             .i3cd000()
@@ -1274,52 +1235,6 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         }
     }
 
-    fn wr_tx_fifo(&mut self, bytes: &[u8]) {
-        let mut chunks = bytes.chunks_exact(4);
-        for chunk in &mut chunks {
-            let word = u32::from_le_bytes([chunk[0], chunk[1], chunk[2], chunk[3]]);
-            self.i3c
-                .i3cd014()
-                .write(|w| unsafe { w.tx_data_port().bits(word) });
-        }
-
-        let rem = chunks.remainder();
-        if !rem.is_empty() {
-            let mut tmp = [0u8; 4];
-            tmp[..rem.len()].copy_from_slice(rem);
-            let word = u32::from_le_bytes(tmp);
-            self.i3c
-                .i3cd014()
-                .write(|w| unsafe { w.tx_data_port().bits(word) });
-        }
-    }
-
-    fn rd_fifo<F>(&mut self, mut read_word: F, out: &mut [u8])
-    where
-        F: FnMut() -> u32,
-    {
-        let mut chunks = out.chunks_exact_mut(4);
-        for chunk in &mut chunks {
-            let val = read_word();
-            chunk.copy_from_slice(&val.to_le_bytes());
-        }
-
-        let rem = chunks.into_remainder();
-        if !rem.is_empty() {
-            let val = read_word();
-            let bytes = val.to_le_bytes();
-            rem.copy_from_slice(&bytes[..rem.len()]);
-        }
-    }
-
-    fn rd_rx_fifo(&mut self, out: &mut [u8]) {
-        self.rd_fifo(|| self.i3c.i3cd014().read().rx_data_port().bits(), out);
-    }
-
-    fn rd_ibi_fifo(&mut self, out: &mut [u8]) {
-        self.rd_fifo(|| self.i3c.i3cd018().read().bits(), out);
-    }
-
     fn ibi_enable(&mut self, config: &mut I3cConfig, addr: u8) -> Result<(), I3cDrvError> {
         let dev_idx = config
             .attached
@@ -1356,6 +1271,13 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
             .modify(|_, w| w.ibithldsignalen().set_bit());
 
         let events = I3C_CCC_EVT_INTR;
+        // ccc_events_set requires HardwareTransfer trait bound on Self.
+        // We are inside HardwareTransfer impl for Ast1060I3c.
+        // Rust might have trouble inferring if Self: HardwareTransfer is not fully established yet?
+        // But Ast1060I3c implements HardwareTransfer (this block).
+        // However, ccc_events_set takes `&mut impl HardwareInterface`.
+        // Ast1060I3c implements HardwareInterface (blanket impl over all sub-traits).
+        // So this call should be valid.
         let _ = ccc_events_set(self, config, dev.dyn_addr, true, events);
 
         i3c_debug!(self.logger, "i3cd030 (SIR reject) = {:#x}", sir_reject);
@@ -1420,10 +1342,16 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         let p = config
             .curr_xfer
             .swap(core::ptr::null_mut(), Ordering::AcqRel);
+        
         if p.is_null() {
-            i3c_debug!(self.logger, "end_xfer: no current xfer");
+            // Drain the response queue to prevent interrupt loops if no xfer is active
+            let nresp = self.i3c.i3cd04c().read().respbufblr().bits() as usize;
+            for _ in 0..nresp {
+                let _ = self.i3c.i3cd010().read().bits();
+            }
             return;
         }
+
         let xfer: &mut I3cXfer = unsafe { &mut *(p.cast::<I3cXfer>()) };
 
         let nresp = self.i3c.i3cd04c().read().respbufblr().bits() as usize;
@@ -1468,7 +1396,12 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
             if err == 0 {
                 if let Some(rx_buf) = cmd.rx.as_deref_mut() {
                     self.rd_rx_fifo(&mut rx_buf[..rx_len]);
+                } else {
+                    // Drain if no buffer provided
+                     self.drain_fifo(|| self.i3c.i3cd014().read().rx_data_port().bits(), rx_len);
                 }
+            } else if rx_len > 0 {
+                 self.drain_fifo(|| self.i3c.i3cd014().read().rx_data_port().bits(), rx_len);
             }
         }
         let mut ret = 0;
@@ -1528,7 +1461,6 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         config: &mut I3cConfig,
         payload: &mut CccPayload<'_, '_>,
     ) -> Result<(), I3cDrvError> {
-        // init i3c_cmd to all 0
         let mut cmds = [I3cCmd {
             cmd_lo: 0,
             cmd_hi: 0,
@@ -1565,7 +1497,6 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
             let cmd = &mut cmds[0];
 
             if id <= 0x7F {
-                // -------- Broadcast CCC --------
                 is_broadcast = true;
 
                 if data_len > 0 {
@@ -1583,7 +1514,6 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
                 else {
                     return Err(I3cDrvError::Invalid);
                 };
-                // -------- Direct CCC --------
                 let pos_ops = config.attached.pos_of_addr(tgt_addr);
                 i3c_debug!(
                     self.logger,
@@ -1863,28 +1793,6 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         }
     }
 
-    fn target_tx_write(&mut self, buf: &[u8]) {
-        self.wr_tx_fifo(buf);
-        let cmd = field_prep(COMMAND_PORT_ATTR, COMMAND_ATTR_SLAVE_DATA_CMD)
-            | field_prep(
-                COMMAND_PORT_ARG_DATA_LEN,
-                u32::try_from(buf.len()).map_or(0, |v| v),
-            )
-            | field_prep(COMMAND_PORT_TID, Tid::TargetRdData as u32);
-
-        self.i3c.i3cd00c().write(|w| unsafe { w.bits(cmd) });
-    }
-
-    fn drain_fifo<F>(&mut self, mut read_word: F, len: usize)
-    where
-        F: FnMut() -> u32,
-    {
-        let nwords = (len + 3) >> 2;
-        for _ in 0..nwords {
-            let _ = read_word();
-        }
-    }
-
     fn handle_ibi_sir(&mut self, config: &mut I3cConfig, addr: u8, len: usize) {
         i3c_debug!(self.logger, "handle_ibi_sir: addr=0x{:02x}", addr);
         let pos = config.attached.pos_of_addr(addr);
@@ -1912,7 +1820,6 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         }
 
         for _ in 0..nibis {
-            // i3c_debug!(self.logger, "i3cd018 (IBI Queue status) = {:#x}", self.i3c.i3cd018().read().bits());
             let reg = self.i3c.i3cd018().read().bits();
 
             let ibi_id = field_get(reg, IBIQ_STATUS_IBI_ID, IBIQ_STATUS_IBI_ID_SHIFT);
@@ -1945,46 +1852,19 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
             }
         }
     }
+}
 
-    fn i3c_aspeed_isr(&mut self, config: &mut I3cConfig) {
-        self.disable_irq();
-        let status = self.i3c.i3cd03c().read().bits();
-        i3c_debug!(self.logger, "[ISR] 0x{:08x}", status);
-        if status == 0 {
-            self.enable_irq();
-            return;
-        }
+impl<I3C: Instance, L: Logger> HardwareTarget for Ast1060I3c<I3C, L> {
+    fn target_tx_write(&mut self, buf: &[u8]) {
+        self.wr_tx_fifo(buf);
+        let cmd = field_prep(COMMAND_PORT_ATTR, COMMAND_ATTR_SLAVE_DATA_CMD)
+            | field_prep(
+                COMMAND_PORT_ARG_DATA_LEN,
+                u32::try_from(buf.len()).map_or(0, |v| v),
+            )
+            | field_prep(COMMAND_PORT_TID, Tid::TargetRdData as u32);
 
-        if config.is_secondary {
-            if status & INTR_DYN_ADDR_ASSGN_STAT != 0 {
-                let da = self.i3c.i3cd004().read().dev_dynamic_addr().bits();
-                if let Some(tc) = &mut config.target_config {
-                    tc.addr = Some(da);
-                }
-                ibi_workq::i3c_ibi_work_enqueue_target_da_assignment(I3C::BUS_NUM.into());
-            }
-
-            if (status & INTR_RESP_READY_STAT) != 0 {
-                self.target_handle_response_ready(config);
-            }
-
-            if (status & INTR_CCC_UPDATED_STAT) != 0 {
-                self.target_handle_ccc_update(config);
-            }
-        } else {
-            if (status & (INTR_RESP_READY_STAT | INTR_TRANSFER_ERR_STAT | INTR_TRANSFER_ABORT_STAT))
-                != 0
-            {
-                self.end_xfer(config);
-            }
-
-            if (status & INTR_IBI_THLD_STAT) != 0 {
-                self.handle_ibis(config);
-            }
-        }
-
-        self.i3c.i3cd03c().write(|w| unsafe { w.bits(status) });
-        self.enable_irq();
+        self.i3c.i3cd00c().write(|w| unsafe { w.bits(cmd) });
     }
 
     fn target_ibi_raise_hj(&self, config: &mut I3cConfig) -> Result<(), I3cDrvError> {
@@ -2136,3 +2016,4 @@ impl<I3C: Instance, L: Logger> HardwareInterface for Ast1060I3c<I3C, L> {
         }
     }
 }
+
