+++ src/main.rs
@@ -346,7 +346,6 @@ fn main() -> ! {
     let mut hace_controller = HaceController::new(hace);
 
     run_hash_tests(&mut uart_controller, &mut hace_controller);
-
     run_hmac_tests(&mut uart_controller, &mut hace_controller);
 
     // Test the owned digest API
